refactor: moved download utilities to their own file
